added shimming layer
