+++ docs/shimming.md
@@ -2,6 +2,8 @@
 
 **A story about configuration**
 
+## Trait path
+
 Steps to handling traits during a "Transition and Deprecation Period".
 
 1. If the argument is prefixed with `ServerApp`, pass this trait to `ServerApp`.
@@ -24,3 +26,44 @@ Steps to handling traits during a "Transition and Deprecation Period".
         2. Migrate/write the trait to a new config file if it came from a config file.
         2. Pass trait to Step 2 above.
     - If the argument is *not* a trait of `ExtensionApp` and not a trait of either `NotebookApp` or `ServerApp`, raise a `"Trait not found."` error.
+
+
+
+## How JupyterApp's Parse Config.
+
+**Valid for traitlets 4.3.x**
+
+Here's a snapshot of how JupyterApp's are initialized. The order of operations we care about goes as follows:
+
+1. `argv` is parsed.
+2. `parse_command_line()` is called and `argv` is passed to this method.
+3. `load_config_file()` is called to load traits from various files.
+
+```python
+# Pulled from Traitlets 4.3.2
+
+    @catch_config_error
+    def initialize(self, argv=None):
+        # don't hook up crash handler before parsing command-line
+        if argv is None:
+            argv = sys.argv[1:]
+        if argv:
+            subc = self._find_subcommand(argv[0])
+            if subc:
+                self.argv = argv
+                self.subcommand = subc
+                return
+        self.parse_command_line(argv)
+        cl_config = deepcopy(self.config)
+        if self._dispatching:
+            return
+        self.migrate_config()
+        self.load_config_file()
+        # enforce cl-opts override configfile opts:
+        self.update_config(cl_config)
+        if allow_insecure_writes:
+            issue_insecure_write_warning()
+```
+
+## Where do we add the steps above?
+
